backend/requirements.txt
djangorestframework
django-cors-headers
psycopg2
requests
schedule
pysftp
redis
python-dotenv
django-redis
openai